Grudge (#16)
This adds the Grudge feature

By enabling this feature ekill remembers killed off elements on each page and tries to get rid of them on subsequent visits to the same pages. By default this feature is disabled and marked as experimental.

An options page is also included, which houses a minimal ui for listing and removing grudges. The chosen widget is a Treeview, (https://github.com/jonmiles/bootstrap-treeview), which depends jQuery, as well as a slightly dated version, v3.3.4, of Bootstrap, (https://github.com/twbs/bootstrap). These assets are also part of this commit.

A small testsuite is included as well, which resides in the `test` subfolder. Tests are meant to run in-browser and are written using Mocha, https://github.com/mochajs/mocha, Chai, https://github.com/chaijs/chai and an extension to Chai for doing deep object comparisons, https://github.com/michelsalib/chai-shallow-deep-equal.

Fixes #9.

Co-authored-by: Duart Snel <mr.codercuracao@gmail.com>